refactor commit local stack fetching
commit-id:7529635e
